Address PR #30 review feedback (Int64 for 32-bit-safe address decoding)
From the Gemini review of fix-forward 4/4: decode integer address sub-fields
as Int64 rather than Int. On 32-bit platforms (watchOS / older devices) Int
is 32-bit, so a large 64-bit value (id / timestamp) would fail Int decoding
and fall through to Double — losing precision / emitting scientific notation.
Int64 decodes losslessly on all architectures.

Tests: add a large-integer (> Int32) address-preservation case.

+++ Sources/SwiftSx/SearchResult.swift
@@ -195,7 +195,10 @@ public struct SearchResult: Codable, Sendable, Equatable {
                     dict[key.stringValue] = value
                 } else if let value = try? nested.decode(Bool.self, forKey: key) {
                     dict[key.stringValue] = String(value)
-                } else if let value = try? nested.decode(Int.self, forKey: key) {
+                } else if let value = try? nested.decode(Int64.self, forKey: key) {
+                    // Int64 (not Int) so large 64-bit values decode losslessly on
+                    // 32-bit platforms instead of falling through to Double (which
+                    // would lose precision / emit scientific notation).
                     dict[key.stringValue] = String(value)
                 } else if let value = try? nested.decode(Double.self, forKey: key) {
                     dict[key.stringValue] = String(value)
